Fixed sprites, added more
The game isn’t running as reliably as I’d like, but I’ve got everything
displaying again and have updated many of the sprites. There are many
new wall sprites: the first 16 have wall parapets following the rule
(side is not also stone) in a binary encoding for
top/bottom/left/right. It’ll draw a nice shaded border around it. Also
there’s a wilderness texture and the water texture which are 2x2 in
size, not currently implemented either.