Fixed up the low-res sprites again; they have their own .png and it has the very specific settings needed to make it not be distorted - we need those pixels to be painted just as they are. This means:
 - Generate Mip Maps must be off
 - Format must not be 'compressed'
 - Filter Mode should be 'Point'
 - Packing Tag should be [RECT]

Strictly speaking the packing tag does not matter much, as most of the sprites are in fact quite rectangular anyway.

Format is the main issue, really, as lossy compression causes the edges of the cells to shift slightly. Unless all sprites are padded with transparent space around their edges, this makes artifacts that look like vertical or horizontal lines appear; those lines are actually the edge of the next sprite.