CastleWallSpriteController
I’ve left all the failed attempts to make sense of the tiles and map
in. At least the tiles are named and organized to make quick work of
this for anyone who understands it. I know how I’d do it but Playfield
doesn’t do anything remotely like that, so I am stymied.