Comments on bark orientation issue
This is where I’m looking at the problem, and it can instantiate a bark
correctly no matter which orientation we are. But, if we flip the
character while a bark is underway, the bark is reversed too. I don’t
know how to find the relevant bark if it exists, and unflip it when
flipping. Removing the bark would also work.