Fixed 'flipped' barks
I hope this is not too much change, but I really did two things: the
flip methods now establish a desired state- the first work out what they
want, then flip if that's not the current stat.

Also, the GetComponents() method lets me get from the CreatureControl to
the BarkControllers on the same object.