Fixed saving & restoring
Fixed the bug where a saved game would not restore; turns out if we
delete a critter upon restoration, we must delete the critter's
inventory.

I think there's still an issue restoring consistently when there are
multiple creatures with the same name; but my intention is to give each
critter a unique name, so this should not be a serious problem.