Camera work
As an experiment, the camera now takes two frames to zip to its goal.
Also, the zoom levels are more hardcoded to directly work with map zoom
levels, and the nearest zoom is nearer. I think this works with the
caves to make them seem bigger, but as long as the zoom levels are the
same relative levels to each other (currently 3-4-5) it ought to be
good.