Science Experiment: Unlock Characters
I’m getting an ‘object reference not set to an instance of an object’
bug, but this is most of the grunt work for this one. It should unlock
characters if you as the player kill them, and if a character wins the
game they re-lock themselves :) ultimate goal, win with all characters
unlocked?