Activateable Characters
The previous effort was doomed because it was in CharacterController,
which isn't used at all during the intro scene, and because it tried to
use the buttons as a data model, when they don't exist during play.

So: I've moved the button wrangling to StartupController, but I've also
created a data model class that can be used at any time and
automatically persists its data to a file.

CreatureController can update this data, and the StartupController can
read it, and it all seems to work.