Fixed Intro Screen Issues
I've reconnected the subtitle and load button, and rearranged slightly
so all elements are more-or-less on screen. The title and load button
are in the lower left now, where there was room. I've also shifted the
character images to the left, so Lady has a better chance to be on
screen.

I've also removed the button prefabs. I guess they were supposed to
provide access to the buttons during game play, but they don't do that.
They don't do anything.