Identified weapon damage design bug
This has been bothering me the whole time, hope I am not putting it too
strongly. If I could fix it, I would do so in an instant, but I can’t
begin to make your weapon item collection class dice roll mechanics
work and would not try to replace it when it’s designed to be the way
it is. Thus, I plead in a comment.