Moved NPCs around
Too many guys on the top level map; I've moved Kim, Jim and Bob to the
caves. We now have 2 player candidates in each area- except there are
none in the castle (they could not get out) and 3 in the caves, but 1 on
the top level.

I'm hoping this will make them not kill each other off so fast.